Updates to volunteer application submission

See merge request campuses/nlr/sep-2025/java-blue/student-pairs/java-blue-finalcapstone-team0!55

+++ java/src/main/java/com/techelevator/model/Applicant.java
@@ -8,7 +8,7 @@ public class Applicant extends ShelterApplication {
 //    first_name VARCHAR(100) NOT NULL,
 //    last_name VARCHAR(100) NOT NULL,
 //    email VARCHAR(150) NOT NULL UNIQUE,
-    private int volunteerApplicationId;
+    private Long volunteerApplicationId;
     @NotBlank
     private String firstName;
     @NotBlank
@@ -29,11 +29,11 @@ public class Applicant extends ShelterApplication {
         this.phoneNumber = phoneNumber;
     }
 
-    public int getVolunteerApplicationId() {
+    public Long getVolunteerApplicationId() {
         return volunteerApplicationId;
     }
 
-    public void setVolunteerApplicationId(int volunteerApplicationId) {
+    public void setVolunteerApplicationId(Long volunteerApplicationId) {
         this.volunteerApplicationId = volunteerApplicationId;
     }
